requirements.txt
python-telegram-bot==20.7
pymongo==4.6.0
python-dotenv==1.0.0
python-dateutil==2.8.2
pandas==2.1.1
python-multipart==0.0.6
